mononoke: use open_filenodes when --skip-caching is set
Summary:
The blob repo creation code was accidentally not calling open_filenodes when
creating a Filenodes DB without caching. This means that if you tried to
connect to filenodes with `--skip-caching` set, you wouldn't go to the sharded
version of this DB.

(this had been the case historically "no caching" used to mean "use SQLite",
but unfortunately I missed this when I added support for no caching + not
SQLite in D15941644 and left the non-sharding implementation in there)

Fortunately, the impact of this should be fairly minimal since we effectively
run everything that matters with caching turned on.

While debugging this, I got tripped up by the `filenodes-shards` CLI argument.
It's unused, so this diff removes it.

Reviewed By: HarveyHunt

Differential Revision: D16894641

fbshipit-source-id: 151c1f5df76b2e5d77214143974912100427817a