help: fix test-check-help.t to recognize fold
Summary:
The latest update in the documentation references the `hg fold` subcommand
from the "core" mercurial files. This makes the `test-check-help.t` test
want to verify the `help` for `fold`. The problem is that the test does not
enable the `amend` extension that provides the fold command.

Following a previous fix to this test, I am enabling the amend extension for
the test. It makes sense to me given the current setup.

Reviewed By: singhsrb

Differential Revision: D16907781

fbshipit-source-id: b0ba31518b6d4c210eb5b8cc7985a945ab1180e4