mononoke: make blobimport use unredacted repo
Summary:
Currently blobimport fails if it tries to import a redacted blob. We should
only enforce redaction in the scenarios where end users could see the data.
Internal imports should be fine to proceed.

As a separate note, problems like this would probably go away when redaction
would be moved to the `FileStore` level.

Reviewed By: krallin

Differential Revision: D16914875

fbshipit-source-id: e531a6248f66b47de87c7aedcbc928bb7ad4ebc9