mononoke: remove Id from ManifestUnodeId
Summary:
We needed this hack because in derive_unode_manifest crate we needed to implement Loadable trait (previously defined in
blobstore crate) for ManifestUnodeId which was defined in mononoke_types crate.
Rust doesn't let us do that - see
https://stackoverflow.com/questions/35160995/implement-foreign-trait-for-foreign-type.

So we used this new Id type, which made it awkward to use. This diff fixes it.

Reviewed By: krallin

Differential Revision: D16885044

fbshipit-source-id: 329b09bb3b671a59824d1d2597a3907bebb82642