treemanifest: prevent fetches that pass the target as the base
Summary:
We've seen problems where we send requests to the server for a manifest
node and claim that we have that same node as the base node. This results in the
server sending no data back. This is obviously a bug in the client, since if
we're trying to fetch the node it means we don't have it and shouldn't use it as
a base. Our current best guess is that the node is in the index but the actual
data is corrupt, so getmissing() claims it exists, but reads actually fail.

Until we figure out the root cause, let's prevent findrecenttree from returning
the node as a base to itself.

Reviewed By: sfilipco

Differential Revision: D16996425

fbshipit-source-id: 04b0cfbe1c432749f2aeb63f98c7b59ab50facc0