manifest: update insert validation
Summary:
The insert code would be unclear in what kind of issue it ran into when
inserting files. Sometimes the file we want to insert is a directory and
other times it want to traverse a directory. This change makes those
situations clear along with some other corner case behaviors.

Reviewed By: quark-zju

Differential Revision: D16775354

fbshipit-source-id: 50ab6bc52b70cc5cef013d11050eb3cdf5b160a5