mononoke: use quicksand_multiplier
Summary:
Previously we had separate load shedding configs for quicksand and
non-quicksand. They were hard to manage because we had to duplicate each load
shedding option for quicksand and non-quicksand. More importantly we had to
duplicate that in the code, and that was error prone.

Now we have a quicksand_multiplier instead which is applied to every
loadshedding option if request is coming from quicksand. let's use it

Reviewed By: krallin

Differential Revision: D16986262

fbshipit-source-id: 46a2ed9eb6e591d8c8a70f8d57efb1cb7836db78