eden: add FUSE_NO_OPENDIR_SUPPORT to capsLabels
Summary:
on linux kernel 5.1 we noticed that some caps were requested but not
known to the diagnostic code.  This little diff adds a missing entry.

There is still one more that is unknown, but I think that we'll need to update
our copy of the fuse header to get that handled here.

Reviewed By: chadaustin

Differential Revision: D16953397

fbshipit-source-id: 30d19843a2cf1ec22fd469654773afaafa8e3b9b