eden: add simple version gate for `eden redirect fixup`
Summary:
We saw some weird cases where we'd get error spew when
folks were using `eden redirect fixup` against really old servers,
despite us having fallback logic for those.

For now let's do a simple manual pre-check; we use july first as
the min server version for this; it was technically available earlier
than that, but this seems like a safe date to use in case of delays
in pushing out that code.

Reviewed By: strager

Differential Revision: D17094330

fbshipit-source-id: 239378906d382cb249a74e1534d0cd363988eaed