enable treemanifest in snapshots
Summary:
The 2018-11 and 2019-03 snapshots were created with flatmanifest
repositories. Flatmanifest support is going away, so update them with
treemanifest repositories.

I untarred them and added an hgrc containing the following contents:

```
[extensions]
treemanifest =

[treemanifest]
treeonly = True

[remotefilelog]
reponame = snapshot
```

It's a bit janky to update a repo in place like this, but it means the
snapshot doesn't have to be recreated from scratch.

Reviewed By: simpkins

Differential Revision: D17103193

fbshipit-source-id: 62da8c2fe579d90281af83d4d8912cc64b09897b