hgmain: fix buildinfo compilation & functionality
Summary:
In D16866464, we changed the name of the `args` to `full_args`, which is
causing compilation to fail when buildinfo is enabled:

https://fburl.com/sandcastle/f23unjvj

That said, buildinfo hasn't actually been working (AFAICT), since argv[0] is
the name of the binary, not the first argument, so running `hg buildinfo`
doesn't work, regardless of whether it was compiled in or not. This diff also
fixes that, and reduces the scope of conditional compilation to reduce the
likelihood of this happening again.

Reviewed By: quark-zju

Differential Revision: D17111929

fbshipit-source-id: 720344721f1fe73a3ebe8330c278e4be83d80c00