eden: fixup `eden redirect fixup` boo-boo
Summary:
I added a None check to make redirect fixup work on macos
where we don't return running build information, but this broke the
integration tests (which don't run on mac!) where the build information
is actually the empty string.

Only check the version date if it is truthy, otherwise it will always
evaluate as less than the cutoff date and short-circuit the fixup
command.

Reviewed By: fanzeyi

Differential Revision: D17113336

fbshipit-source-id: 38dff9157a760b747958722e70567871dccd8cc0