manifest: update finalize to work on durable trees
Summary:
In some cases the finalize algorithm is used to persist data that is received
in a bundle. The process is that it constructs a store from the bundle and
goes to construct a tree with the root node received. It then goes through
finalize to generate the entries that need to be written to local storage.

Reviewed By: quark-zju

Differential Revision: D17125149

fbshipit-source-id: de5a1e922a6aebe48e238d8473177a8d3f7a9ef5