revisionstore: handle indexedlog corruption
Summary:
We've seen a handful of cases where the indexedlogdatastore becomes corrupted
which makes Mercurial unable to run properly. For now, and since we only use
the indexedlog for the hgcache, let's just remove it.

A better solution would be to harden the indexedlog code to better detect the
corruption and attempt to fix them.

Reviewed By: quark-zju

Differential Revision: D17115622

fbshipit-source-id: ee474a5df60c4414f6ea21ace7dff0f7048879c9