Back out "hg: wrap convert in a big transaction for better pack behavior"
Summary:
I broke some svn tests in that commit, but while fixing forward
realized that we're blocked from making this change because we rely on
rolling back the last commit in the case where the filemap creates an
empty commit.

We can revisit this when we have convert this all to rust!

Reviewed By: sfilipco

Differential Revision: D17140172

fbshipit-source-id: c93911f599afd2d22e84b3327a2a9455204065a0