remotefilelog: add a test to show incremental local data repack
Summary:
This illustrate that the local data repack isn't a full one. Note that the test
is a bit flaky due to 1) the underterministic nature of what folder to repack,
and 2) the choice of folder to repack being done twice (one for data, one for
history). Since I won't land this without the next one, this is probably fine.

Reviewed By: quark-zju

Differential Revision: D17135978

fbshipit-source-id: 641d257d0e308b2bb4d91fdc4c214b22e45b4911