mononoke: fix fastlog batching for empty commits
Summary:
I ran into a tricky case while deriving FastlogBatch for empty commits:
* If a commit is empty then new unode manifest is not created [1]
* But FastlogBatch's derived data logic was still creating a new FastlogBatch.
This batch would point not to the commit that created the unode, but to the
latest empty commit.

This is obviously not a huge deal because it affects only root entry only for
empty commits. But it can be confusing and tricky to investigate, so I think
it's better to fix it.

[1] There is one exception - empty commit with no parents

Reviewed By: krallin

Differential Revision: D17160654

fbshipit-source-id: 949d35860efe2f3df6ec1c11b4c6c4d077982c95