revisionstore: always try to rescan when calling get_missing
Summary:
When created, the PackStore are empty as no disk scan are performed. As
get_missing is one of the first API to be called on the store when a prefetch
is needed, that means that get_missing always pretend that everything is
missing, which isn't true.

Since we don't want to always rescan on get_missing, let's respect the scanning
frequency and then figure out what's missing.

Reviewed By: quark-zju

Differential Revision: D16794077

fbshipit-source-id: 460fda8f118c1b36e5d5b29472e8a06ef0754ec9