mononoke: find modified unodes in merge commits
Summary:
This is the first step in supporting merges in apiserver's fastlog.

To generate FastlogBatches we need to find unodes that were modified.
For a commit with single parent it's just a diff between this commit and it's
parent.

That would work for merges as well, however it might do unnecessary work.
For example, let's say first parent introduced file A, second parent introduced
file B. Diff between merge commit and first parent would return file B, and
we'd try to generate FastlogBatch for it even though this FastlogBatch is
already generated. That might be a problem in case of big merges.

To avoid this problem let's filter out unodes that exist in any of the parents.

Reviewed By: krallin

Differential Revision: D17180324

fbshipit-source-id: ecfa6748937159fdcdab9c7f08d964230f87739a