third-party: initial import https://github.com/serde-rs/bench
Summary:
I'd like to have a mincode-like serialization format that is really simple:
vlqencoding for integers and use bincode-like strategy for the rest.

There is a mincode crate, but it does not compile due to extra depenencies and
out-of-date serde. It has extra features like bitpack, bitvec which are less
interesing in our commit serialization usecase.

The serde API needs a lot of boilerplate since it supports 20+ types. To make
it easier, I'm importing dtolnay's [serde-rs/bench](https://github.com/serde-rs/bench/commits/9718508c304d3b930ff0e034eec0cc96e42d6505)
as a starting point.

The serde API needs a lot of boilerplate. Import a minimal implementation so we
can customize on top of it. serde-rs/bench is similar to bincode but much smaller:

bincode:

   333 config.rs
   486 de/mod.rs
   185 de/read.rs
   115 error.rs
   173 internal.rs
   173 lib.rs
   788 ser/mod.rs
  2253 total

mincode:

    57 bitpack.rs
    84 bitvec.rs
    95 float.rs
    90 lib.rs
   434 refbox.rs
   103 rustc_serialize/mod.rs
   427 rustc_serialize/reader.rs
   460 rustc_serialize/writer.rs
   123 serde/mod.rs
   661 serde/reader.rs
   682 serde/writer.rs
  3216 total

serde-rs/bench:

  407 de.rs
   60 error.rs
   25 lib.rs
  442 ser.rs
  934 total

The following (unrelated) files are excluded in this import:

  README.md
  benches/bincode.rs
  tests/bincode.rs
  .gitignore
  .travis.yml

The code does not compile internally due to mismatched `bincode` version. It
will be fixed in the next diff.

Reviewed By: dtolnay

Differential Revision: D17087353

fbshipit-source-id: 26329b36e1488c0c6149287f1f2dcd89acd15b0b