tp2/rust-crates-io: Update structopt to remove no_version
Summary:
When we upgraded from structopt 0.2 to 0.3 in D17138630, we had to put a no_version on all of our structopt data structures or else structopt would fail to compile with the error:

```
`CARGO_PKG_VERSION` environment variable is not defined, use `version = "version"` to
set it manually or `no_version` to not set it at all
```

Fbcode isn't using Cargo so there wasn't a CARGO_PKG_VERSION associated with each cli.

I raised with the maintainers in [TeXitoi/structopt#243](https://github.com/TeXitoi/structopt/issues/243) that requiring no_version attributes everywhere is annoying, and it's been fixed in structopt 0.3.1. This diff updates to 0.3.1 and removes 42 no_version attributes from fbcode.

Reviewed By: bolinfest

Differential Revision: D17231406

fbshipit-source-id: 2bab2864afbf23b34ea5d73884462455d863c139