mononoke: fix the build
Summary:
I had a diff that removed `safe_writes` that landed at about the same time
ikostia's diff that added a new command that includes `safe_writes`, so the
`mononoke-buck` build is broken as a result. This fixes that.

I also noticed a build warning (unused future) while checking that the build
was fixed, so I also fixed that.

Reviewed By: ikostia, StanislavGlebik

Differential Revision: D17257765

fbshipit-source-id: 47ac7e46b5263878f8aa7fd3e3326e356fecbbe6