mononoke/filestore: implicitly fuse ChunkStream
Summary:
I noticed while testing the LFS proxy that we could occasionally poll the
stream underlying a `ChunkStream` when it had already indicated it had no more
data. This is bad because it might panic on some streams.

I initially thought this might have to do with the `poll()` shenanigans we do
in `prepare()`, but that doesn't appear to be the case (getting rid of those
doesn't fix anything).

Regardless, we can fix this by implicitly fusing things in `ChunkStream`, which
seems to be the problem here (in the LFS server, uploads that don't get chunked
don't crash, which makes sense we only use stdlib combinators to concatenate
the incoming stream).

Reviewed By: StanislavGlebik

Differential Revision: D17225985

fbshipit-source-id: 76ad21744461cdbd1376c6ce810e0a7a10b37ef0