mononoke/lfs_server: add a LFS server for Mononoke
Summary:
This introduces a LFS server for Mononoke, which is designed to also act as a
reverse proxy to a fallback LFS server. The goal of this LFS server is to
uphold the following guarantees:

- If a client uploads successfully, then the LFS content will be present in
  Mononoke as well as the upstream server.
- If a piece of content is present in either Mononoke or the upstream server,
  then downloading this content will succeed. If the content is available in
  Mononoke, it'll be served there. Otherwise, the client will be redirected to
  the upstream server.

Note that while the LFS server for Mononoke exposes a route per repo, that
isn't the case for the upstream server.

Also, note that this only does proxying for the batch endpoint and uploads.
Clients must have access to the upstream to download from there blobs that
don't exist in Mononoke.

Implementation-wise, this uses async Rust and the Gotham web framework (Gotham
is based on Hyper, which we also use to talk to upstream LFS). Note that
neither Gotham nor Mononoke's filestore doesn't use Tokio 0.2 yet, so we still
have to run on Tokio 0.1 as the executor.

Reviewed By: StanislavGlebik

Differential Revision: D17182753

fbshipit-source-id: 24c9c9c7183506c38d7534665a19d42acf4b3442