apiserver: create new Mononoke alongside apiserver Mononoke
Summary:
Instantiate the new Mononoke internal API's `Mononoke` object alongside the
existing apiserver `Mononoke` abstraction.  We will migrate all uses of the
existing one to the new one and then remove the apiserver's `Mononoke`.

For now, construct the new one by assembling it from the parts of the existing
`Mononoke` abstraction.  When that is removed, we will create the new one
directly.

Provide the new Mononoke object to the HTTP server and the Source Control
Service Thrift server, as both of these will need it.

Reviewed By: krallin

Differential Revision: D16984061

fbshipit-source-id: eb8c237dfa6d82a96d4cb0baf29e4bfa39119bc9