mononoke_api: create new Mononoke internal API
Summary:
Start the creation of a new Mononoke internal API within the `mononoke_api` crate.

This abstraction will be used by the API server, as well as any future server types.

To create a new Mononoke instance the caller instantiates a `Mononoke` object,
passing it the configuration to load.

Queries begin by calling a method on the `Mononoke` instance, passing a
`CoreContext` for that request.  For example, calling the `repo` method returns
a `RepoContext` object which is bound to the repository and the request
`CoreContext`.  Further methods on these context objects can be called, which
will return either new data or new, more specialized, context objects.

For example, to find the author of a commit, the caller can use:
```
    let ctx = CoreContext::new(...);
    let author = mononoke.repo(ctx, "reponame").changeset(changeset_specifier).author();
```

Reviewed By: StanislavGlebik

Differential Revision: D16803206

fbshipit-source-id: 20a8e08f0075476844df4b8c5015ac3faa22d423