mononoke: support merges in FastlogBatch
Summary:
Finally support merges. Whenever we have a merge a completely new batch is
created.

The one thing that surprised me was that ParentOffsets actually can be
negative. For example
```
     o
  /    \
 o       o
 |       |
 o       |
 |       |
 o <---  | --- ParentOffset of this commit can be negative!
   \   /
     o
```

It happens because of BFS order of traversal - parent can be visited befor the
child. Note that negative offsets shouldn't cause problems.

Reviewed By: aslpavel

Differential Revision: D17183355

fbshipit-source-id: b5165ffef7212ce220dd338079db9e26a3030f58