update chg to correctly honor the pager.stderr setting
Summary:
Update the chg code to correctly honor the `pager.stderr` setting, and avoid
piping stderr to the pager when it is disabled.

Since only the server-side code knows the hg config values, the server-side
chgserver.py code passes this config setting back to the client when sending
the pager request.

Reviewed By: quark-zju

Differential Revision: D17109106

fbshipit-source-id: 6b69b1a7de9f61db51af7b0ba00d65fa5053a795