refactor the chg code that sends system and pager requests
Summary:
This refactors the code the code that sends system and pager requests from the
chg server to the client.

Previously these were both sent using the `S` channel code.  The `S` channel
code was slightly unusual and had special handling: in general the code
assumed that upper-case channel codes did not include any request body data,
but this wasn't true for the `S` channel data.  I changed this to lower-case
in order to eliminate this special case handling, and I also split up the
system and pager data into different channel codes, since they have fairly
different behavior.  System requests are now sent with an `s` channel code,
and pager requests are sent with a `p` channel code.

I also changed the code to require that the server always adds a terminating
`\0` byte after each environment variable value.  Previously the client code
was responsible for adding a nul terminator on the last string, which could
potentially require the client to copy the data into a larger buffer in order
to do so.

I also made a minor change to the client-side `readchannel()` code so that it
can read the channel type and body data length with a single system call
instead of making 2 separate `recv()` calls.

The main benefit of these changes is that they will let the server pass some
additional configuration information with pager requests.  The change to make
use of this new field will come in a subsequent diff.

Reviewed By: quark-zju

Differential Revision: D17216291

fbshipit-source-id: c3044cf3d5f5e103f0b62d083e4ef3764160f20e